Converting to modular code.
Issue with the main and UI class working together.
